Refactors KubnernetesInstances class to be non static, introduces KubernetesInstanceFactory that creates KubernetesInstance instances.
