+++ src/main/java/cd/go/contrib/elasticagent/KubernetesAgentInstances.java
@@ -38,25 +38,25 @@ public class KubernetesAgentInstances implements AgentInstances<KubernetesInstan
     public Clock clock = Clock.DEFAULT;
     private boolean refreshed;
     private KubernetesClientFactory factory;
+    private KubernetesInstanceFactory kubernetesInstanceFactory;
 
     public KubernetesAgentInstances() {
-        this(KubernetesClientFactory.instance());
+        this(KubernetesClientFactory.instance(), new KubernetesInstanceFactory());
     }
 
     public KubernetesAgentInstances(KubernetesClientFactory factory) {
+        this(factory, new KubernetesInstanceFactory());
+    }
+
+    public KubernetesAgentInstances(KubernetesClientFactory factory, KubernetesInstanceFactory kubernetesInstanceFactory) {
         this.factory = factory;
+        this.kubernetesInstanceFactory = kubernetesInstanceFactory;
     }
 
     @Override
     public KubernetesInstance create(CreateAgentRequest request, PluginSettings settings, PluginRequest pluginRequest) throws Exception {
         KubernetesClient client = factory.kubernetes(settings);
-        KubernetesInstance instance;
-        if(isUsingPodYaml(request)) {
-            instance = KubernetesInstance.createUsingPodYaml(request, settings, client, pluginRequest);
-        } else {
-            instance = KubernetesInstance.create(request, settings, client, pluginRequest);
-        }
-
+        KubernetesInstance instance = kubernetesInstanceFactory.create(request, settings, client, pluginRequest, isUsingPodYaml(request));
         register(instance);
 
         return instance;
@@ -119,7 +119,7 @@ public class KubernetesAgentInstances implements AgentInstances<KubernetesInstan
                 Map<String, String> podLabels = pod.getMetadata().getLabels();
                 if (podLabels != null) {
                     if (StringUtils.equals(Constants.KUBERNETES_POD_KIND_LABEL_VALUE, podLabels.get(Constants.KUBERNETES_POD_KIND_LABEL_KEY))) {
-                        register(KubernetesInstance.fromInstanceInfo(pod));
+                        register(kubernetesInstanceFactory.fromKubernetesPod(pod));
                     }
                 }
             }
@@ -155,9 +155,13 @@ public class KubernetesAgentInstances implements AgentInstances<KubernetesInstan
             DateTime dateTimeCreated = new DateTime(createdAt);
 
             if (clock.now().isAfter(dateTimeCreated.plus(period))) {
-                unregisteredInstances.register(KubernetesInstance.fromInstanceInfo(pod));
+                unregisteredInstances.register(kubernetesInstanceFactory.fromKubernetesPod(pod));
             }
         }
         return unregisteredInstances;
+    }
+
+    public boolean instanceExists(KubernetesInstance instance) {
+        return instances.contains(instance);
     }
 }
